refactor(app): typed CreatorVote, manifest-only mod.rs, tighter visibility
